# QL_FOODDELI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QL_FOODDELI API (Dapper & SignalR)",
    "description": "Hệ thống backend API quản lý giao đồ ăn nhanh QL_FOODDELI - Sprint 5",
    "version": "v1"
  },
  "paths": {
...
    "/api/CuaHang/{id}": {
      "get": {
        "tags": [
          "CuaHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "CuaHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CuaHang"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CuaHang"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CuaHang"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CuaHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/MonAn/search": {
      "post": {
        "tags": [
          "MonAn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAnSearchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAnSearchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonAnSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MonAn": {
      "get": {
        "tags": [
          "MonAn"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "MonAn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MonAn/{id}": {
      "get": {
        "tags": [
          "MonAn"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "MonAn"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonAn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "MonAn"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MonAn/CuaHang/{maCuaHang}": {
      "get": {
        "tags": [
          "MonAn"
        ],
        "parameters": [
          {
            "name": "maCuaHang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/NguoiDung/{id}": {
      "get": {
        "tags": [
          "NguoiDung"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/NguoiDung/Email": {
      "get": {
        "tags": [
          "NguoiDung"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Voucher/kiem-tra": {
      "post": {
        "tags": [
          "Voucher"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KiemTraVoucherRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KiemTraVoucherRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KiemTraVoucherRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Voucher": {
      "post": {
        "tags": [
          "Voucher"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoucherCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VoucherCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VoucherCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CuaHang": {
        "type": "object",
        "properties": {
          "maCuaHang": {
            "type": "string",
            "nullable": true
          },
          "maChuShop": {
            "type": "string",
            "nullable": true
          },
          "tenCuaHang": {
            "type": "string",
            "nullable": true
          },
          "moTa": {
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "type": "string",
            "nullable": true
          },
          "diaChi": {
            "type": "string",
            "nullable": true
          },
          "anhCuaHang": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KiemTraVoucherRequest": {
        "type": "object",
        "properties": {
          "maVoucher": {
            "type": "string",
            "nullable": true
          },
          "tongTienDonHang": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "MonAn": {
        "type": "object",
        "properties": {
          "maMonAn": {
            "type": "string",
            "nullable": true
          },
          "maCuaHang": {
            "type": "string",
            "nullable": true
          },
          "maDanhMuc": {
            "type": "string",
            "nullable": true
          },
          "tenMonAn": {
            "type": "string",
            "nullable": true
          },
          "moTa": {
            "type": "string",
            "nullable": true
          },
          "gia": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "anhMonAn": {
            "type": "string",
            "nullable": true
          },
          "danhGia": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trangThai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recordCount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MonAnSearchRequest": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "tenMonAn": {
            "type": "string",
            "nullable": true
          },
          "maDanhMuc": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VoucherCreateRequest": {
        "type": "object",
        "properties": {
          "maVoucher": {
            "type": "string",
            "nullable": true
          },
          "tenVoucher": {
            "type": "string",
            "nullable": true
          },
          "loaiGiamGia": {
            "type": "string",
            "nullable": true
          },
          "giaTri": {
            "type": "number",
            "format": "double"
          },
          "giamToiDa": {
            "type": "number",
            "format": "double"
          },
          "giaTriDonToiThieu": {
            "type": "number",
            "format": "double"
          },
          "ngayBatDau": {
            "type": "string",
            "format": "date-time"
          },
          "ngayKetThuc": {
            "type": "string",
            "format": "date-time"
          },
          "soLuong": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Dán mã JWT Token thu được sau khi Đăng nhập vào ô này (KHÔNG gõ chữ Bearer):",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}